Optionally Hide Debug Output (Finished Games Only)
It would be desirable to hide all debug output the engine does internally to give the use complete control of what gets printed and what doesn't. 

Intended for complete games. Should never be used otherwise so bugs can be debugged.

HIDDEN_DEBUG_OUTPUT needs to be defined by a checkbox in LateralGM/RadialGM and a flag in emake for these changes to even be usable.