[skip ci] use same title bar icon for game+dialogs
WIP. looking for a way to fix a few things to get this working before we can merge it.

Intended outcome: copies the icon used by window_set_icon() to dialog windows, use xlib default (generic) icon when no icon is set.
What we currently have: dialog icon gets destroyed and replaced with fully transparent/invisible icon. No idea why.

I don't like duplicating the XSetIconFromSprite function over from the xlib platform into our widget code, but there's not much of a better option when we want this to also be used in SDL but want to avoid SDL depending on xlib when it happens to using Wayland depending on the current session it is running on (when not using widgets).