Add function initialize_working_directory() (#1674)
* Update CocoaMain.cpp

This code was tested in the GameMaker equivalent with the help of my File Manager extension and it is verified working.

This function will set the working directory to the app bundle's Resources folder like GM4Mac 7.5, GMStudio 1.4, GMS 2.x and most Mac apps do, if the executable is in an app bundle. If the executable is not in an app bundle, use the executable path.

ONLY use working_directory for loading read-only included files! When SAVING, use game_save_id. Mac app bundles must be read-only, otherwise your app will not work anymore after being code-signed because you were stupid enough to tamper with it. Your app will also get rejected from the Mac App Store if you do that. We still need to add game_save_id because not only is it the GM sandbox directory, Mac Apps require sandboxing as an operating system requirement, unless you distribute outside their store without code-signing.

* removed useless code

removed code that doesn't take into account the possible lacking of an application bundle, which is necessary when testing your games regardless of how you distribute your app.