Begin DragonFlyBSD Support. (#2243)
Most of the work was already done when we added FreeBSD Support. DragonFlyBSD is a fork of an old version of FreeBSD so there doesn't need much to be changed other than mainly an additional platform check for everything.

I did this with the git command line btw. @JoshDreamland you did say once I learned that you could make me a member of enigma-dev. Please consider doing that since I've been capable of doing this and knew how for a while now. I can do it more often now that I'm more confident with it and don't have the same hangups I used to have.

The only thing that doesn't work right now is LGM because we need a version of JNA that has the DragonFlyBSD native library inside it. I tried using the one from the JNA DragonFlyBSD package and it still didn't recognize it. We can worry about that later, or simply wait until RGM is ready and forget about supporting LGM entirely with this.