Fix the 'ouble' bug
This bug occurs when the `pos` variable in `collect_variables()`
(`collect_variables.cpp:57`) is set one character too far after erasing
an initializer. This leads to the output getting mangled because of the
skipped character.

For example, consider the script
```
local double i;
local double j;
```

This will lead to the output being
```
double i;
ouble j;
```

Because after parsing `i`, the `pos` will be on the second character of
`local`, i.e. it will see `ocal`, and the way the parser is coded makes
it so that it ends up skipping the first `d` of `double`.

To fix this, just set `pos` to be one character behind when erasing an
initializer.