Fix ending parenthesis of macro being considered as a macro argument
When given a macro like MACRO(a, b), this tokenises to 'MACRO', '(', 'a', ',', 'b', ')'. The ending ')' is not considered as a closing parenthesis i.e. as closing off the macro, thus it is inserted as an argument by the macro expansion system. This commit fixes that by making it so that macro arguments are only inserted before the closing parenthesis is encountered which sets the 'paren' variable to 0