Fix incorrect macro argument stringification in the lexer
This is caused by 'errc.snippet' being used in place of the actual string synthesized to be used. The reason this is done is because the lexer may need to refer to the stringified form of the macro argument long after this is done however the string storing the stringified form is scoped to the function it is created in, thus it will not live as long as the lexer and can cause issues when being referred to after it is destructed.

To fix this, introduce a set to store stringified forms of macro arguments for as long as the lexer lives by adding it as a member variable of the lexer. This will ensure that stringified forms of macro arguments never cause lifetime issues and will also handle deduplicating stringified forms of arguments which stringify to the same string.