WIP: Fixes for macros in EDL lexer (#2305)
* Add test for macro expansion

The const_cast is used to insert the macro itself forcefully into the lexer so that it can be expanded without having to invoke the preprocessor

* Fix ending parenthesis of macro being considered as a macro argument

When given a macro like MACRO(a, b), this tokenises to 'MACRO', '(', 'a', ',', 'b', ')'. The ending ')' is not considered as a closing parenthesis i.e. as closing off the macro, thus it is inserted as an argument by the macro expansion system. This commit fixes that by making it so that macro arguments are only inserted before the closing parenthesis is encountered which sets the 'paren' variable to 0

* Fix incorrect macro argument stringification in the lexer

This is caused by 'errc.snippet' being used in place of the actual string synthesized to be used. The reason this is done is because the lexer may need to refer to the stringified form of the macro argument long after this is done however the string storing the stringified form is scoped to the function it is created in, thus it will not live as long as the lexer and can cause issues when being referred to after it is destructed.

To fix this, introduce a set to store stringified forms of macro arguments for as long as the lexer lives by adding it as a member variable of the lexer. This will ensure that stringified forms of macro arguments never cause lifetime issues and will also handle deduplicating stringified forms of arguments which stringify to the same string.