Add "System Information " Extension
Should be fully implemented correctly on Windows, macOS, and Linux. FreeBSD and other platforms will need more stuff implemented and are not completely finished. Haven't tested the STIGMA version of this. It was originally a GameMaker extension which I verified worked as such, but it might have careless mistakes since the time I converted it to a STIGMA compliant extension. We'll see. If so, I'll be quick to fix the mess I've made, as usual.